+++ upyutils/rsa/pem.py
@@ -97,3 +97,27 @@ def load_pem(contents, pem_marker):
     # Base64-decode the contents
     pem = b''.join(pem_lines)
     return ubinascii.a2b_base64(pem)
+
+
+def save_pem(contents, pem_marker):
+    """Saves a PEM file.
+    :param contents: the contents to encode in PEM format
+    :param pem_marker: the marker of the PEM content, such as 'RSA PRIVATE KEY'
+        when your file has '-----BEGIN RSA PRIVATE KEY-----' and
+        '-----END RSA PRIVATE KEY-----' markers.
+    :return: the base64-encoded content between the start and end markers, as bytes.
+    """
+
+    (pem_start, pem_end) = _markers(pem_marker)
+
+    b64 = ubinascii.b2a_base64(contents).replace(b'\n', b'')
+    pem_lines = [pem_start]
+
+    for block_start in range(0, len(b64), 64):
+        block = b64[block_start:block_start + 64]
+        pem_lines.append(block)
+
+    pem_lines.append(pem_end)
+    pem_lines.append(b'')
+
+    return b'\n'.join(pem_lines)